[MISC] Use sharg

+++ include/shared.hpp
@@ -1,5 +1,5 @@
-// SPDX-FileCopyrightText: 2006-2024 Knut Reinert & Freie Universität Berlin
-// SPDX-FileCopyrightText: 2016-2024 Knut Reinert & MPI für molekulare Genetik
+// SPDX-FileCopyrightText: 2006-2025 Knut Reinert & Freie Universität Berlin
+// SPDX-FileCopyrightText: 2016-2025 Knut Reinert & MPI für molekulare Genetik
 // SPDX-License-Identifier: BSD-3-Clause
 
 #pragma once
@@ -22,7 +22,7 @@ static inline constexpr uint64_t adjust_seed(uint8_t const kmer_size,
 struct all_arguments
 {
     std::filesystem::path path_out{"./"};
-    uint8_t threads{1};
+    uint16_t threads{1u};
 };
 
 //!\brief arguments used for estimate, ibf, minimiser
